# rest-api/resources/src/main/resources/openapi/jobTarget/job-scopeId-jobId-targets.yaml
openapi: 3.0.3

info:
  title: Eclipse Kapua REST API - Jobs Targets
  version: '1.0'
  contact:
    name: Eclipse Kapua Dev Team
    url: https://eclipse.org/kapua
    email: kapua-dev@eclipse.org
  license:
    name: Eclipse Public License 2.0
    url: https://www.eclipse.org/legal/epl-2.0

paths:
  /{scopeId}/jobs/{jobId}/targets:
    get:
      tags:
        - Job - Target
      summary: Get all the Jobs Targets
      operationId: jobTargetList
      parameters:
        - $ref: '../openapi.yaml#/components/parameters/scopeId'
        - $ref: '../job/job.yaml#/components/parameters/jobId'
        - $ref: '../openapi.yaml#/components/parameters/sortParam'
        - $ref: '../openapi.yaml#/components/parameters/sortDir'
        - $ref: '../openapi.yaml#/components/parameters/limit'
        - $ref: '../openapi.yaml#/components/parameters/offset'
        - $ref: '../openapi.yaml#/components/parameters/askTotalCount'
      responses:
        200:
          description: The list of the Targets of the Job
          content:
            application/json:
              schema:
                $ref: './jobTarget.yaml#/components/schemas/jobTargetListResult'
        400:
          $ref: '../openapi.yaml#/components/responses/illegalArgument'
        401:
          $ref: '../openapi.yaml#/components/responses/unauthenticated'
        403:
          $ref: '../openapi.yaml#/components/responses/subjectUnauthorized'
        500:
          $ref: '../openapi.yaml#/components/responses/kapuaError'
    post:
      tags:
        - Job - Target
      summary: Create a new Job Target
      operationId: jobTargetCreate
      parameters:
        - $ref: '../openapi.yaml#/components/parameters/scopeId'
        - $ref: '../job/job.yaml#/components/parameters/jobId'
      requestBody:
        description: An object containing the properties for the new Job Target to be created
        content:
          application/json:
            schema:
              $ref: '../jobTarget/jobTarget.yaml#/components/schemas/jobTargetCreator'
        required: true
      responses:
        201:
          description: The Job Target that has just been created
          content:
            application/json:
              schema:
                $ref: '../job/job.yaml#/components/schemas/job'
        400:
          $ref: '../openapi.yaml#/components/responses/illegalArgument'
        401:
          $ref: '../openapi.yaml#/components/responses/unauthenticated'
        403:
          $ref: '../openapi.yaml#/components/responses/subjectUnauthorized'
        404:
          $ref: '../openapi.yaml#/components/responses/entityNotFound'
        500:
          $ref: '../openapi.yaml#/components/responses/kapuaError'